Refactor NoteBlock
